Update: ProductServiceClientPactTest : Allow testSingleProduct to be run in isolation
The current test configures the mockserver URI only for getAllProducts. 
getSingleProduct works because the base URL is already set.
It is better to allow the methods to work in isolation, so that they for example can run individually in an IDE.
This fix is applied only on the step3 branch, but other branches will likely have the same issue.